Restructure into internal packages and cmd/corediff binary
Extract shared code from flat cmd/ package into internal/:
- internal/hashdb: HashDB type with Load/Save/Contains/Add
- internal/normalize: line normalization, hashing, file extension checks
- internal/path: app root detection, path exclusion patterns
- internal/highlight: suspect pattern detection

New CLI at cmd/corediff/ with subcommands:
- scan (default): unchanged scan behavior
- db add: hash files into DB (fixed broken call signature)
- db merge: combine databases (now functional)

Single file scanning skips app root check automatically.
