+++ internal/highlight/highlight_test.go
@@ -1,11 +1,11 @@
-package main
+package highlight
 
 import "testing"
 
 var dummy = []byte("lorum ipsum fopen lorum ipsum fopen lorum ipsum fopen lorum ipsum fopen lorum ipsum fopen ")
 
-func BenchmarkHighlight(b *testing.B) {
+func BenchmarkShouldHighlight(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		shouldHighlight(dummy)
+		ShouldHighlight(dummy)
 	}
 }
